# encoding/json: optimize tagOptions.Contains