The new code is about 18% faster.

Here is the benchmark highlights:

goos: windows
goarch: amd64
pkg: encoding/json
cpu: Intel(R) Core(TM) i5-10200H CPU @ 2.40GHz
                     │  old.bench  │              new.bench              │
                     │   sec/op    │   sec/op     vs base                │
TagOptionsContains-8   17.07n ± 1%   13.87n ± 3%  -18.78% (p=0.000 n=10)